+++ stelvio/rich_deployment_model.py
@@ -110,7 +110,6 @@ RESOURCE_TYPE_NAMES: dict[str, str] = {
     "aws:docdb/subnetGroup:SubnetGroup": "DocumentDB Subnet Group",
     "aws:docdb/clusterParameterGroup:ClusterParameterGroup": "DocumentDB Parameter Group",
     "aws:secretsmanager/secretRotation:SecretRotation": "Secret Rotation",
-    "pulumi-python:dynamic/aws:DocumentDbSecretRotationDisabled": "DocumentDB Rotation Disable",
 }